+++ src/Formatting.Analyzers/CSharp/DiagnosticIdentifiers.Generated.cs
@@ -60,5 +60,6 @@ namespace Roslynator.Formatting.CSharp
         public const string NormalizeWhitespaceAtEndOfFile = "RCS0058";
         public const string PlaceNewLineAfterOrBeforeNullConditionalOperator = "RCS0059";
         public const string BlankLineAfterFileScopedNamespaceDeclaration = "RCS0060";
+        public const string BlankLineBetweenSwitchSections = "RCS0061";
     }
 }